Refactor : 예외 중복부분 추상화 적용
