+++ src/main/java/com/flab/delivery/exception/SessionLoginException.java
@@ -2,16 +2,9 @@ package com.flab.delivery.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class SessionLoginException extends RuntimeException {
-
-    HttpStatus httpStatus;
+public class SessionLoginException extends AbstractException {
 
     public SessionLoginException(String message, HttpStatus httpStatus) {
-        super(message);
-        this.httpStatus = httpStatus;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return this.httpStatus;
+        super(message, httpStatus);
     }
 }
